move to_energy to units
