+++ src/Model.jl
@@ -1,20 +1,6 @@
 # Contains the physical specification of the model
 
 using Unitful
-import Unitful.uconvert
-using UnitfulAtomic
-
-# Convert a temperature to an energy via the Boltzmann constant.
-function uconvert(a::Unitful.EnergyUnits, x::Unitful.Quantity{T,Unitful.𝚯,U}) where {T <: Real, U <: Unitful.Units}
-    uconvert(a, Unitful.k * x)
-end
-
-# Convert a value to an energy in atomic units.
-# If given a number, assume it is already in atomic units (pass through directly).
-function to_energy(x::Unitful.Quantity)::Real
-    ustrip(uconvert(UnitfulAtomic.Eh_au, x))
-end
-to_energy(x::Real)::Real = x
 
 # A physical specification of a model.
 # Contains the geometry information, but no discretization parameters.
